[BACKPORT 2024.1][PLAT-13638] Change error message for reverse direction xCluster
Summary:
**Context**
During failover/switchover, YBA will create an xCluster config in the reverse direction.
If there are extra or missing tables from the databases under replication on the DR replica,
then YBA will refuse to proceed with the failover/switchover.
Prior to the changes in this diff, the error message only indicated that not all tables in the namespace were selected.
The user may need extra help understanding why this is the case and what they should do to address this issue.

**Change**
The reverse direction xCluster precheck is only run for failover and switchover.
Thus, we can offer a more specific error message here which tells the user how to address the issue.

Further, we only show the extra tableIds in the error message and not every tableId in the namespace.
This change helps the user pick out exactly which tables they need to look at. Every other table is already present in both DR repica and DR config.

Original commit: 237f95ba200807645849b2baae842a9b090fed8f / D35904

Test Plan:
Create a DR config.
- Add extra tables on the DR replica.
- Drop tables on the DR replica
- Verify that error message appears as intended.
{F257743}

Reviewers: hzare, cwang, vbansal, sanketh

Reviewed By: hzare

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36059