[#22810] CDCSDK: Use slot name to distinguish between consumption models
Summary:
With the introduction of replication slot consumption model, there is a requirement to distinguish between the two consumption models internally in the code. Till now this was being done by the flag `FLAGS_ysql_TEST_enable_replication_slot_consumption` which is set to true only for replication slot consumption model. However, with GA, this flag will be made an auto flag with default value true and hence can no longer be used to distinguish between the two consumption models. Additionally, a given database (namespace) can have streams belonging to one of the two consumption models.

This diff introduces changes to use replication slot name present in the stream metadata to distinguish between the two consumption models. The distinction was required in three broad areas. Firstly, the behaviour on creation of dynamic tables is different in replication slot consumption and hence distinction between the models is needed there. Secondly, the Update Peers and Metrics has an optimisation in place to move the history retention barriers forward on tables not in publication list. This optimisation requirres distnction between the two models. Lastly, we need the distinction between the two models to decide whether to use replica identity or CDC Recordtype for the before image.

All the tests in the `CDCSDKConsumptionConsistentChangesTest` were previously using a stream which was created with `CreateConsistentSnapshotStream` test method, which internally called the CreateCDCStream rpc. However, this caused the streams to not have a slot name associated with them. In this diff CreateConsistentSnapshotStream has been replaced by `CreateConsistentSnapshotStreamWithReplicationSlot` in CDCSDKConsumptionConsistentChangesTest for all the tests. This creates the streams using a replication connection and the streams have a slot name associated with them.
Jira: DB-11708

Test Plan:
Existing CDC Tests
./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestCreationOfSlotOnNewDBAfterUpgrade

Reviewers: asrinivasan, stiwary, skumar, siddharth.shah

Reviewed By: asrinivasan

Subscribers: ybase, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35781