[BACKPORT 2024.1][PLAT-14433] Use correct kubeconfig for edit provider validation
Summary:
Original diff and commit: D36051 0ada80aa6d8da8e84ab4e6eb6bddbba250eda4d1

During edit provider, kubeConfigContent is not present in the
KubernetesInfo, so we defaulted to using in-cluster credentials
instead of using the kubeconfig which provider has.

With this change, we will check if kubeConfigContent is present, if
not, check kubeConfig and use it. If both are not present we will
default to using in-cluster credentials.

Also, fixes an edge case where issuer is provided with non-existing
namespace.

Test Plan:
- Removed in-cluster RBAC of YBA pod and then tried to edit an invalid
  provider. It used the kubeconfig from provider and validated all the
  parameters.
- Tried adding a bogus kubeconfig, it failed as expected.
- Also tried to create new provider, it is getting validated as
  expected.

Reviewers: svarshney, rohita.payideti, #yba-api-review!, anijhawan

Reviewed By: svarshney

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36177