[PLAT-14443] YBA Installer wait for ready time configurable.
Summary:
We should be able to configure how long yba installer will wait for
yba to be "ready" after an install or upgrade.

I also improved how we set the current schema version to not require devs to update a constant.

Test Plan: tested install and an install + upgrade with the new confi goption

Reviewers: muthu, sanketh

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36067