[PLAT-14325] Reduce wait time for PITR config creation from 15 seconds to 1 second
Summary:
This diff refactors the CreatePITRConfig code and ensures all the check steps have retrying with runtime modifiable poll delay and timeout. With the new poll delay default value the PITR creation time reduces to about 2 seconds.

It also removes the unused class RestorePitrConfig.

It overloads the method `doWithModifyingTimeout` so that it can accepts a Runnable and properly throw the exception of the main function when retrying times out.

Test Plan:
- Made sure PITR is set up in about 2 seconds.
- Made sure when the delay is 100ms, the retry logic happens a couple of times before the create PITR task completes.
- Made sure when the timeout is 1s, the proper exception is thrown.

Reviewers: sanketh, cwang, vbansal, vpatibandla, #yba-api-review

Reviewed By: sanketh, cwang, #yba-api-review

Subscribers: sanketh, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36035