[BACKPORT 2024.1][#22748] YSQL: Support for 'yboutput' plugin
Summary:
Original commit: 2a000837150c37006507f1a6fe46c70508bf039f / D35932
This diff introduces support for `yboutput` output plugin. Since the code for `yboutput` and the `pgoutput` is same to a great extent, inorder to prevent code duplication a new file for `yboutput` is not created. Instead, at the time of loading the output plugin, we make a decision based on the plugin name.

A callback `yb_support_yb_specifc_replica_identity_cb` has been introduced in this revision and all the output plugins need to implement this. For the test_decodig and wal2json plugins this callback has a NOOP implementation.  This callback is called at the time of loading the output plugin. While loading the `yboutput` plugin, the callback implementation initializes the variable `is_yboutput_mode` to true which is used in handling the changes specific to `yboutput`.

This diff also adds checks to ensure that tables with unsupported replica identity (CHANGE) are not present in the streaming list.

######Backport description
No merge conflicts were encountered.

Jira: DB-11651

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testReplicaIdentityChangeShouldNotWorkWithPgoutput'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testDynamicTableWithReplicaIdentityChangeWithPgoutput'

Reviewers: stiwary, asrinivasan, skumar, siddharth.shah

Reviewed By: skumar

Subscribers: ycdcxcluster, yql, jason

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36077