[pg15] refactor: remove unnecessary slot to tuple conversion (part 2)
Summary:
This revision is in continuation with a0bc5efb474b33ffefaea7a7fcd7ba3571564fed to reduce the dependency of YB code on HeapTuple and instead work with TupleTableSlot.

This revision refactors the function YBCGetYBTupleIdFromTuple() to take TupleTableSlot as an argument instead of HeapTuple and rename it to YBCComputeYBTupleIdFromSlot(). This removes the need to convert slot to tuple by the callers (except for YBCExecuteUpdateLoginAttempts(), which has HeapTuple and now needs to convert it to TupleTableSlot before calling YBCGetYBTupleIdFromTuple()).

Consequently,
- the function YBCApplyInsertRow() no longer sets ybctid on the tuple. Introduce a heap tuple-based wrapper `YBCExecuteInsertHeapTupleForDb` on `YBCExecuteInsertForDb` that copies ybctid from slot to heap tuple.
- convert slot to heap tuple before calling CacheInvalidateHeapTuple() for catalog relations.

Test Plan: Jenkins: rebase: pg15

Reviewers: jason, amartsinchyk

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36001