[BACKPORT 2024.1][#22894] CDCSDK: Support creation of an old model stream in an upgraded environment
Summary:
Original commit: c020ace23a04022b60b20d13908280d4262b688a / D35978
This diff introduces changes to support creation of an old model stream via yb-admin in an upgraded environment. With the upgrade to `2024.1.1` all the test and preview flags guarding the replication slot consumption feature will be turned into default true autoflags. In order to support creation of an old model stream in such an upgraded environment we need to take care of the decisions which are presently being made on the basis of these flags in stream creation path. This diff handles this by adding a check on the presence of replication slot name in the request in the stream creation path.

This diff also adds a check in the stream creation path to ensure that in a given database only one kind of streams can be present. If any user tries to create a replication slot on a namespace with old model stream already present (or vice versa), the stream creation will fail.

######Backport Description
No merge conflicts were encountered.

Jira: DB-11796

Test Plan:
Existing CDC Tests
./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestCreationOfOldStreamAfterUpgrade
./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestFailureCreatingStreamsOfDifferentTypesOnSameDB

Reviewers: asrinivasan, stiwary, skumar

Reviewed By: asrinivasan

Subscribers: ycdcxcluster, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36086