[PLAT-13367] change localProvider key from task to input
Summary:
changes localProvider key from task to input. Task keys are not available during sbt tasks
as these are internal keys.
Have changed the same to inputKeys - so that they are visible as part of `sbt tasks`

`taskKey` is used to define a key for a task in sbt.
`inputKey` is used to define a key for an input task in sbt.

Test Plan: iTest pipeline

Reviewers: yshchetinin, sanketh

Reviewed By: yshchetinin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36072