[pg15-upgrade] test: disable expression pushdown during upgrade
Summary:
Vanilla PG11 and PG15 generate different pushdown expressions for some operations, and process some pushed-down operations differently. It's currently a WIP to make PG15 compatible with PG11 while the cluster is in mixed mode during the upgrade. One notable casualty at present is that running \d on a PG11 tserver fails with the error:

  ERROR:  could not determine which collation to use for string comparison

What's happening is this: YB masters are restarted as YB PG15 binaries early in the upgrade process. The \d command sends queries for PG catalogs, which are hosted on the YB master. One of these queries pushes down an invocation of the namene function. This function requires a collation to be specified in PG15, but PG11, which doesn't need a collation for this function, doesn't send one.

When we implement expression pushdown compatibility in PG15, to handle this case we will translate from the PG11 namene invocation to one that's compatible with the PG15 internals. It's likely we will simply specify the C collation, to match the PG11 behavior. Until we implement this translation, to avoid hitting known problems whenever pushdown occurs, all binaries for both PG11 and PG15 should be run with yb_enable_expression_pushdown set to false throughout the upgrade.

Change the upgrade tests to specify this setting throughout the upgrade. Additionally, add a test for \d, which now works in mixed mode as a result.

Test Plan:
  ./yb_build.sh release --sj
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh

Reviewers: jason

Reviewed By: jason

Subscribers: swapnil.shivam, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D36100