[pg15] test: fix regress test yb_pg_stat_backend
Summary:
In a pg_stat_activity output, one of the queries changes to "GRANT ALL
ON SCHEMA public TO public".  This is likely related to YB pg15
50f46b2474bf89dfdce14c5302b049d899ac603a, so simply update the expected
output.  This fixes the entire schedule, so TestPgRegressPgStat should
now pass for the most part.  On master and pg15 branches, it is flaky
because the first two beids can sometimes flip.

Test Plan:
On Almalinux 8, manually check this passes >20%:

    ./yb_build.sh fastdebug --gcc11 --java-test org.yb.pgsql.TestPgRegressPgStat --tp 1 -n 50

Jenkins: rebase: pg15, test regex: TestPgRegressPgStat

Reviewers: aagrawal, fizaa, tfoucher

Reviewed By: fizaa

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36104