[pg15] fix: assertion for foreign index build
Summary:
In the index build code, there is a YB-introduced assertion that fails
on foreign relations.  This was an oversight.  Fix the assertion
criteria.

The value of the local variable "concurrent" shouldn't matter for
foreign tables since foreign tables can't be partitioned (there is an
error message that depends on that) and indexes can't be created on
foreign tables.

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --java-test TestPgRegressForeignData

should no longer crash.  It still fails for other reasons.

Jenkins: rebase: pg15

Reviewers: fizaa, tfoucher

Reviewed By: fizaa

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36154