[#22807] docdb: Add gflag to control of RWCLock debug logging
Summary:
The debug logging for slow locks in `rwc_lock.cc` is currently only available when compiling in debug mode. This diff changes it to be guarded by a new gflag, `enable_rwc_lock_debugging`. This allows us to turn on the flag as needed for debugging.

This diff also fixes a bug where we would access a `CowObject`'s state even though we don't have a read lock or write lock on the object. The `state()` code does have a line that was supposed to guard against this: ` DCHECK(lock_.HasReaders() || lock_.HasWriteLock());`. But:
- In debug builds, `HasWriteLock` was only checking that the thread id was the same as the one that owns the lock (and not checking whether the write lock was held)
- In non-debug builds, DCHECK will not cause a crash
Jira: DB-11705

Test Plan: Turned on the flag and ran the diff through Jenkins to make sure nothing breaks. Also verified via some local testing that the stack traces still work

Reviewers: hsunder, xCluster

Reviewed By: hsunder

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D35738