[PLAT-14519] Master failover must not fail when some tservers are down
Summary:
The changes are
1. Master failover does not fail even if multiple tservers are down.
2. Check for live tservers are moved to using the existing ybclient API.
3. Skip unreachable tservers while choosing the new master.

Test Plan:
1. Created universe with RF3, 5 nodes in 2 Azs.
2. Stopped one master and another tserver.
3. Verified that the master failover completed fine.
4. Started one tserver. Nothing happened.
5. Started the 2nd tserver.
6. syncMasterAddresses started after a few mins.

Also verified by bringing down a master and a tserver in the same AZ to make findReplacementMaster not return any live tserver. In this case, master failover did not run. Then, when the tserver (leaving the old master stopped) was started, master failover also started.

Reviewers: cwang, sanketh

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36233