[PLAT-14478] Allowing prometheus auth without enabling HTTPS
Summary: Had a bug in our setting of yaml values where if the file was not a valid yaml then we could not set any values. In the enableAuth case for prometheus the web.yml is only generated by config if enableHTTPS is true, wherewas we set the prometheus auth users and passwords after generating the password in go code. This diff improves the SetYamlValue function to also handle the case where the initial file can be empty and we still set the new keys.

Test Plan:
change test to not having any values initially, ensure test fails before change
enable prometheus auth without HTTPS, ensure reconfigure succeeds

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36213