[BACKPORT pg15-cherrypicks][#22937] docdb: Backward scans: make pggate be aware of fast backward scan capability
Summary:
This is a first part of CBO-related changes for backward scans, which is not directly related to
the cost calculation. But the aim of the change is to make pggate be aware of fast backward scan
capability to be able to correctly identify the cost of backward scans in the CBO.
Refer to https://github.com/yugabyte/yugabyte-db/issues/22370 for the details.

For this purpose the flag is moved to common_flags.cc and YBCPgGFlagsAccessor is updated
to be able to get the value of the flag. Additionally the flag is mark as `NON_RUNTIME` to prevent
any confusion from a user side as PG flags does not support runtime change. Such change is
acceptable as it is not expected the flag change in future.

**Upgrade/Rollback safety:**
In the worst case, when the node is not yet upgraded, the planner/optimizer will not be aware of
the fast backward scan capability and as a result would be choosing a different execution path
without fast backward scan optimization avoiding possible performance gain. On the other hand,
if the planner/optimizer already aware of the capability (and it's on), but remote yb-tserver is
not yet upgraded, then it may result in a slower query but with no correctness impact.

**Note.** Merge conflicts.
Original conflicts have been resolved by https://github.com/yugabyte/yugabyte-db/commit/259a71e746946b68c36a9ceeb3e18b235f91205f

New conflicts are due to git merge was not able to handle a new field at the end of a structure, which was
last updated by YB pg15 fedbdac1c20d8301b63860041835f001bfb33118:
- ybc_pg_typedefs.h:397
```
  const bool*     TEST_ysql_hide_catalog_version_increment_log;
<<<<<<< HEAD
  const bool*     TEST_generate_ybrowid_sequentially;
=======
  const bool*     ysql_use_fast_backward_scan;
>>>>>>> 1773ae2545 ([#22937] docdb: Backward scans: make pggate be aware of fast backward scan capability)
} YBCPgGFlagsAccessor;
```

- ybc_pggate.cc:1927
```
          &FLAGS_TEST_ysql_hide_catalog_version_increment_log,
<<<<<<< HEAD
      .TEST_generate_ybrowid_sequentially =
          &FLAGS_TEST_generate_ybrowid_sequentially,
=======
      .ysql_use_fast_backward_scan = &FLAGS_use_fast_backward_scan,
>>>>>>> 1773ae2545 ([#22937] docdb: Backward scans: make pggate be aware of fast backward scan capability)
  };
```

Resolved them manually.

Jira: DB-11853

Original commit: 1773ae2545 / D36245

Test Plan: Jenkins

Reviewers: jason, tfoucher

Reviewed By: jason

Subscribers: yql, ybase, hsunder

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36653