[PLAT-14173] UserIntentOverrides into v2 UniverseSpec
Summary:
The previous diffs D33904, D34466, D34782, D35023 introduced the Universe Spec and Info and the corresponding CRUD operations. This change is to complete the implementation of all properties of the Universe resource.

The main additions are the equivalent of `UserIntentOverrides` into v2. Defined the `ClusterNodeSpec` to hold the properties:
```
ClusterNodeSpec:
    - instanceType
    - storageSpec (equivalent of deviceInfo)
    - cgroupSize
    - master (equivalent of userIntentOverrides.perProcess.master)
        - instanceType
        - storageSpec
    - tserver (equivalent of userIntentOverrides.perProcess.tserver)
        - instanceType
        - storageSpec
    - azNodeSpec (equivalent of userIntentOverrides.azOverrides)
        (will contain the same properties as that of ClusterNodeSpec)
```
And `ClusterNetworkingSpec` with similar nested structure to hold the proxyConfig.

Test Plan: To run `use_yba_v2=true` itest with this change once it lands.

Reviewers: #yba-api-review!, dshubin

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36203