[#23052] yugabyted:  Restarting a node fails when data_dir is missing in user specified configuration.
Summary:
When restarting yugabyted with a user configuration file that does not contain the `data_dir` parameter, an AttributeError is raised because the code attempts to call split on a `NoneType` object.
To fix the issue we simply check if the parameter is present in the given user configuration file, if it is not present then we do not split the parameter.
Jira: DB-11983

Test Plan: Manual Testing.

Reviewers: sgarg-yb, nikhil

Reviewed By: sgarg-yb, nikhil

Subscribers: shikhar.sahay

Differential Revision: https://phorge.dev.yugabyte.com/D36241