[#22840] DocDB: Fix DFATAL in WaitQueue::Impl::HandleWaiterStatusRpcResponse
Summary:
https://github.com/yugabyte/yugabyte-db/commit/2896530b3a9eb209ecb79d4665c0ffb4b10bd52a refactored the way we maintain waiters at the wait-queue, changing `waiter_status_` to track `WaitingTxn` which internally stores all waiting rpc requests, `WaiterData`(s), corresponding to the transaction at the tablet.

Earlier, we used to remove the waiter from `waiter_status_` when executing its callback, but now we only erase the waiter either on processing the txn status response or when we see that all rpc requests of this waiter have been resolved.

the following could happen
```
1. Poll sees a waiting request for the `WaitingTxn`, and launches a status request
2. The blocker for the rpc request resolves, and the waiting rpc request's callback is executed
3. The waiter txn (which is no longer blocked) commits.
4. The txn coordinator processes the status response and returns COMMITTED
5. HandleWaiterStatusRpcResponse processes the response and DFATALs seeing COMMITTED status
```

With the current state of the code, it should be totally okay to see `COMMITTED` or other resolved txn statuses for a `WaitingTxn` which is no longer blocked. Hence DFATAL only if we see resolved txn statuses when `WaitingTxn` has active waiting requests.

Note: This isn't an issue in release mode anyways since the next iteration of `Poll` would purge the `WaitingTxn`.
Jira: DB-11737

Test Plan:
./yb_build.sh fastdebug --cxx-test pgwrapper_pg_wait_on_conflict-test --gtest_filter PgWaitQueueContentionStressTest.TestResumeWaitersOnScheduler -n 50 --tp 1
./yb_build.sh fastdebug --cxx-test pgwrapper_pg_wait_on_conflict-test --gtest_filter PgWaitQueuesTest.DeadlockResolvesYoungestTxn -n 50 --tp 1

Reviewers: pjain, rthallam, patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D36143