[#21500] YSQL: Skip schema version mismatch tests with Connection Manager enabled
Summary:
Certain tests fail with the `schema version mismatch` error when Connection Manager is used as the endpoint. All the tests being modified in this patch will be skipped when Connection Manager is used as the endpoint, they are all dependent on the use of a new physical connection for each logical connection that is spawned during the test runs. Tests modified:

- org.yb.pgsql.TestAlterTableWithConcurrentTxn.testDmlTransactionAfterAlterOnCurrentResourceWithoutCachedMetadata
- org.yb.pgsql.TestAlterTableWithConcurrentTxn.testMultipleDmlTransactionWithAlterOnCurrentResource
- org.yb.pgsql.TestYbBackup.testGeoPartitioningRestoringIntoExisting
- org.yb.pgsql.TestYbBackup.testGeoPartitioningRestoringIntoExistingWithTablespaces

Jira: DB-10384

Test Plan:
Jenkins: enable connection manager, all tests

Verify that the above skipped tests do not fail, and are indeed being skipped

Reviewers: asrinivasan, skumar

Reviewed By: asrinivasan

Differential Revision: https://phorge.dev.yugabyte.com/D36234