[#21550] YSQL: testSetMemoryTracking UT fails with connection manager enabled
Summary:
The test `org.yb.pgsql.TestYbPgStatActivity#testSetMemoryTracking` is failing is while doing the check `total_heap_released>0` . Made a change in the code where a table is created and dropped which would result in the heap memory released. With this change, the test is succeeding with connection manager enabled.
Jira: DB-10435

Test Plan:
Check that UT is passing now with connection manager enabled and disabled.

```
./yb_build.sh --enable-ysql-conn-mgr-test --java-test 'org.yb.ysqlconnmgr.TestYbPgStatActivity#testSetMemoryTracking'
```

Jenkins: all tests, enable connection manager

Reviewers: asrinivasan, rbarigidad, mkumar, skumar, nkumar, telgersma

Reviewed By: rbarigidad, telgersma

Differential Revision: https://phorge.dev.yugabyte.com/D36270