[pg15] test: clean up tests using sequential ybrowid
Summary:
Sequential ybrowid generation makes previous workarounds no longer
necessary.  Update some tests to have less workarounds, enabling the
TEST_generate_ybrowid_sequentially where it can be useful.  In case of
TestPgRegressDml, split it up to TestPgRegressDml and TestPgRegressPgDml
so that pg ported tests are in one schedule with the flag enabled and yb
tests are in a different schedule.

- Remove ORDER BY where unnecessary
- Remove setting VERBOSITY terse because the ordering of cascade details
  is now fixed, even though it often does not match the upstream PG
  ordering due to system table updates that do not reallocate ybrowid.
  This reveals some other deficiencies of not dropping objects in
  yb_pg_alter_generic.

There are more tests that could be adjusted, but stop here for now.

Test Plan:
On Almalinux 8:

    #!/usr/bin/env bash
    set -eu
    ./yb_build.sh fastdebug --gcc11
    pg15_tests/run_tests.sh <<EOT
    JAVA	org.yb.pgsql.TestPgRegressDml
    JAVA	org.yb.pgsql.TestPgRegressPgDml
    JAVA	org.yb.pgsql.TestPgRegressPgMisc
    SHELL	test_pg_misc_independent
    EOT

Jenkins: rebase: pg15, test regex: TestPgRegress

Reviewers: aagrawal, fizaa, tfoucher

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36303