[pg15] test: set compute_query_id=regress globally
Summary:
A number of regress tests set compute_query_id to regress at the top of
the file so that EXPLAIN VERBOSE doesn't output query id.  This is not a
problem with upstream PG because they don't load pg_stat_statements by
default, but for YB, we want to exercise pg_stat_statements, so the
workaround is to instead use compute_query_id.  However, it is tedious
to set it for every single test file that needs it, so instead, set it
globally for all regress tests.

Test Plan:
Depends on D36306
Jenkins: rebase: pg15, test regex: TestPgRegress

Reviewers: aagrawal, fizaa, tfoucher

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36307