[PLAT-14535][PLAT-14149] Azure provider validation fixes
Summary:
The azure library we use for validation is inconsistent with the exceptions and the  error messages that it throws.
This causes tests to fail since we are highlighting fields bases on these error messages.

Made changes to catch general exception instead of specific ones and other minor improvements.

Test Plan: Manually tested all the changes.

Reviewers: #yba-api-review!, svarshney

Reviewed By: svarshney

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36312