[pg15] test: add appendToYsqlPgConf for Java tests
Summary:
Since issue #22711 is still open, add a workaround for Java tests to be
able to append to ysql_pg_conf_csv.  This is needed for future work to
set yb_use_hash_splitting_by_default to false for ported regress tests
and to set compute_query_id to regress for all regress tests.  The
method of implementation I chose does not prevent a developer from
forgetting to use this helper and accidentally overwriting the entire
ysql_pg_conf_csv.  Given the current framework, it is difficult to do
that.

Test Plan:
Depends on D36303
Jenkins: rebase: pg15

Reviewers: aagrawal, fizaa, tfoucher

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36306